fix typo

+++ config/grayScaleImage.h
@@ -2,6 +2,6 @@
 #ifndef SUM_H
 #define SUM_H
 
-int grayScaleImage(int w, int h);
+float grayScaleImage(int w, int h);
 
 #endif
