+++ grayScaleImage.c
@@ -4,7 +4,7 @@
 //   h: height of the image
 // Return value
 //   grayScale size of the image Bitmap (in byte)
-int grayScaleImage(int w, int h) {
+float grayScaleImage(int w, int h) {
    // YOUR CODE HERE - BEGIN
 
    // YOUR CODE HERE - END
